extract some helper methods
