+++ ios/instruments/devicestate.go
@@ -3,7 +3,6 @@ package instruments
 import (
 	"github.com/danielpaulus/go-ios/ios"
 	dtx "github.com/danielpaulus/go-ios/ios/dtx_codec"
-	log "github.com/sirupsen/logrus"
 )
 
 const conditionInducerChannelName = "com.apple.instruments.server.services.ConditionInducer"
@@ -14,13 +13,9 @@ type DeviceStateControl struct {
 }
 
 func NewDeviceStateControl(device ios.DeviceEntry) (*DeviceStateControl, error) {
-	dtxConn, err := dtx.NewConnection(device, serviceName)
+	dtxConn, err := connectInstruments(device)
 	if err != nil {
-		log.Debugf("Failed connecting to %s, trying %s", serviceName, serviceNameiOS14)
-		dtxConn, err = dtx.NewConnection(device, serviceNameiOS14)
-		if err != nil {
-			return nil, err
-		}
+		return nil, err
 	}
 	conditionInducerChannel := dtxConn.RequestChannelIdentifier(conditionInducerChannelName, loggingDispatcher{dtxConn})
 	return &DeviceStateControl{controlChannel: conditionInducerChannel, conn: dtxConn}, nil
